Refactor DiscoveryService and TimelineGenerator into discovery package
Modularized the monolithic apts/discovery.py into a structured package:
- Moved DiscoveryService to apts/discovery/service.py.
- Moved TimelineGenerator to apts/discovery/timeline.py.
- Preserved public API via apts/discovery/__init__.py.
- Decomposed DiscoveryService.get_top_picks into focused private helper methods to improve readability and reduce cyclomatic complexity.
- Updated relative imports to account for the new package structure.
- Compiled translations to ensure test suite passes.

This change adheres to the Single Responsibility Principle and improves maintainability of the discovery logic.
