+++ apts/discovery/__init__.py
@@ -1,0 +1,4 @@
+from .service import DiscoveryService
+from .timeline import TimelineGenerator
+
+__all__ = ["DiscoveryService", "TimelineGenerator"]
